recover yolo world

+++ models/siammask.py
@@ -71,11 +71,11 @@ class SiamMask:
         outputs = self.model.run(
             ["output", "target_pos1", "target_sz1", "z_features1", "delta_yx"],
             {
-                "im": im1.astype(np.float32),
-                "target_pos0": target_pos0.astype(np.float32),
-                "target_sz0": target_sz0.astype(np.float32),
+                "im": im1.astype(float),
+                "target_pos0": target_pos0.astype(float),
+                "target_sz0": target_sz0.astype(float),
                 "scale_x": np.array(100, dtype=np.float64),
-                "z_features0": self.dummy_z_feat.astype(np.float32),
+                "z_features0": self.dummy_z_feat.astype(float),
                 "first_time": np.array(True),
             },
         )
@@ -118,9 +118,9 @@ class SiamMask:
         outputs = self.model.run(
             ["output", "target_pos1", "target_sz1", "z_features1", "delta_yx"],
             {
-                "im": im1.astype(np.float32),
-                "target_pos0": self.target_pos.astype(np.float32),
-                "target_sz0": self.target_sz.astype(np.float32),
+                "im": im1.astype(float),
+                "target_pos0": self.target_pos.astype(float),
+                "target_sz0": self.target_sz.astype(float),
                 "scale_x": np.array(scale_x, dtype=np.float64),
                 "z_features0": self.z_feature,
                 "first_time": np.array(False),
@@ -152,7 +152,7 @@ class SiamMask:
         b = (im.shape[0] - 1) / back_box[3]
         c = -a * back_box[0]
         d = -b * back_box[1]
-        mapping = np.array([[a, 0, c], [0, b, d]]).astype(np.float)
+        mapping = np.array([[a, 0, c], [0, b, d]]).astype(float)
         crop = cv2.warpAffine(
             mask,
             mapping,
@@ -161,7 +161,7 @@ class SiamMask:
             borderMode=cv2.BORDER_CONSTANT,
             borderValue=-1,
         )
-        target_mask = (crop > self.segmentation_thres).astype(np.uint8)
+        target_mask = (crop > self.segmentation_thres).astype(int)
         return target_mask
 
     def __preprocess__(self, im, pos, model_sz, original_sz, avg_chans) -> np.ndarray:
